# [Enhancement] Imporve random exchange performance