## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

now:
all chilren will send data to union node, the optimize point
1. union node always use max parallelism node which from chilren
3. union node just rename his chilren data, none other work
4. data need serialized/deserialized, and transform by network

<img width="396" alt="image" src="https://user-images.githubusercontent.com/5609319/173985304-94ee0461-4e0c-4379-8029-b1663d86a73c.png">


Enhancement
1. recompute union node host, always keep same with his chilren
2. update RANDOM choose channel strategy, only choose local node channel

<img width="396" alt="image" src="https://user-images.githubusercontent.com/5609319/173985318-38e30bb8-fd69-41b1-bc40-86b325aa713f.png">


test tpcds query02
before + pass_through: 8s ~ 11.213s,not stable
after: 9.803s
after + pass_through: 3.4s ~ 7.719s