@Seaven 2022-06-21
run starrocks_fe_unittest