@wanpengfei-git 2022-06-21
#### [FE PR Coverage check]

:disappointed: **fail** : 11 / 14 (78.57%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/qe/Coordinator.java|8|11|72.73%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|3|3|100.00%|